+++ src/reduxStore/store.ts
@@ -1,6 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { createSubscription } from 'react-redux/lib/utils/Subscription';
 import { setAccount, setAccountNonce } from 'reduxStore/slices';
+import { loginSessionMiddleware } from './middlewares/loginSessionMiddleware';
 import { persistIgnoredActions, persistStore, reducers } from './persistConfig';
 import rootReducer from './reducers';
 
@@ -21,7 +22,7 @@ export const store = configureStore({
           'providers.accountProvider'
         ]
       }
-    })
+    }).concat(loginSessionMiddleware)
 });
 
 export const subscription = createSubscription(store);