+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[v4.3.3](https://github.com/multiversx/mx-sdk-dapp/pull/1592)] - 2025-09-27
+
+- [Enhance logout handling in useNativeAuthLogout and improve token expiration checks](https://github.com/multiversx/mx-sdk-dapp/pull/1592)
+
 ## [[v4.3.2](https://github.com/multiversx/mx-sdk-dapp/pull/1590)] - 2025-09-26
 
 - [Removed login session expiry and moved guardian storage to session](https://github.com/multiversx/mx-sdk-dapp/pull/1590)